Flatten components
