+++ frameworks/keyed/react-livestore/index.html
@@ -11,7 +11,7 @@
   <script>
     // Redirect to add a trailing slash if none is present
     if (!window.location.pathname.endsWith('/')) {
-      window.location.href = window.location.href.replace(/index\.html$/, '') + '/';
+      // window.location.href = window.location.href.replace(/index\.html$/, '') + '';
     }
   </script>
 </body>
