--- frameworks/keyed/react-livestore/src/components/Button.tsx
@@ -1,16 +1,0 @@
-import React from 'react'
-
-const Button = ({ id, cb, title }) => (
-    <div className="col-sm-6 smallpad">
-        <button
-            type="button"
-            className="btn btn-primary btn-block"
-            id={id}
-            onClick={cb}
-        >
-            {title}
-        </button>
-    </div>
-)
-
-export default Button;
